refactor(runtime): split orchestrator/tests.rs — extract subagent tests
Convert the 3536-line `crates/runtime/src/orchestrator/tests.rs` into
a directory module (`tests/mod.rs` + sibling files) and extract the 10
subagent integration tests into a new `tests/subagent.rs`:

- `subagent_spawn_complete_round_trip`
- `conversation_can_delegate_to_anonymous_subagent`
- `conversation_can_delegate_to_existing_agent_context`
- `subagent_nesting_depth_limit_enforced`
- `subagent_tool_filtering_restricts_tools`
- `subagent_cancellation_stops_loop`
- `subagent_llm_error_records_failed_status`
- `subagent_thinking_step_persisted_to_db` (was nested under
  "Thinking persistence tests")
- `subagent_max_iterations_returns_failed_status` (was under
  "Max-iterations error tests")
- `subagent_forwards_inner_events_to_parent_sink` (was under
  "Thinking event streaming tests")

Helpers (`build`, `mount_answer`, `ollama_*`, `MockExtTool`) remain in
`tests/mod.rs` and are reachable from sibling files via `super::*`.

First slice of a stacked split — more sections to follow (history,
end_turn, sanitize, bus/worker, streaming, multimodal, misc).

cargo test -p assistant-runtime --lib reports 220 passed (same as
before); orchestrator::tests subtotal still 75.
